refactor: share snake case param lookup
(cherry picked from commit 77d2f9a3547f3fad86d3a423d366b9adf3ba35cb)
